requirements.txt
plotly==4.9.0
dash==1.14.0
Pillow==7.2.0
autopep8==1.5.3
scipy<=1.41
numpy<=1.18
pyserial==3.4
pandas<=0.25.3
omegaconf<=1.4.1 
wheel<=0.34.2 
